# json: Improve performance of `php_json_encode_array()`